pyproject.toml
[project]
name = "xythrion-bot"
version = "0.1.0"
description = "Bot portion of Xythrion"
authors = [{ name = "Xithrius", email = "xithrius@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "httpx>=0.27.0",
    "pillow>=10.4.0",
    "numpy>=2.0.1",
    "lxml>=5.2.2",
    "beautifulsoup4>=4.12.3",
    "tabulate>=0.9.0",
    "humanize>=4.10.0",
    "seaborn>=0.13.2",
    "scipy>=1.14.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
]

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
  "pre-commit>=3.8.0",
  "ruff>=0.6.7",
  "basedpyright>=1.17.5",
  "pytest>=8.3.2",
  "anyio>=4.4.0",
  "pytest-cov>=5.0.0",
  "pytest-env>=1.1.4",
  "pytest-xdist>=3.6.1",
]

[tool.rye.scripts]
start = "python3 -m bot"

precommit = "pre-commit install"

# Database migrating
history = "alembic history"
migrate = "alembic revision --autogenerate -m"
upgrade = "alembic upgrade head"
downgrade = "alembic downgrade"

# Testing
test = "pytest --ff"
atest = "pytest -n auto --ff"
vtest = "pytest -vvv --ff"
retest = "pytest --lf"

# Coverage
test-cov = "pytest --cov=./ --cov-report=xml"
html = "coverage html"
report = "coverage report --sort=Cover"

"format:ruff" = "ruff format"
"check:ruff" = "ruff check ."
"check:ruff:ci" = "ruff check --output-format=github ."
"fix:ruff" = "ruff check --fix ."

typecheck = "basedpyright"

format = { chain = ["format:ruff", "fix:ruff"] }
lint = { chain = ["check:ruff", "typecheck"] }
ci = { chain = ["check:ruff:ci", "typecheck"] }

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
xfail_strict = true
filterwarnings = ["ignore::DeprecationWarning"]

# https://docs.basedpyright.com/#/configuration
[tool.pyright]
exclude = [".venv", ".cache", ".ruff_cache", ".pytest_cache"]
pythonVersion = "3.12"

typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false

[tool.ruff]
target-version = "py312"
extend-exclude = [".cache"]
output-format = "concise"
line-length = 120
unsafe-fixes = true
preview = true

[tool.ruff.lint]
select = [
  "ANN",
  "B",
  "C4",
  "D",
  "DTZ",
  "E",
  "F",
  "I",
  "INT",
  "N",
  "PGH",
  "PIE",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "T20",
  "TID",
  "UP",
  "W",
]
ignore = [
  "ANN002",
  "ANN003",
  "ANN401",
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
  "D203",
  "D212",
  "D214",
  "D215",
  "D301",
  "D400",
  "D401",
  "D402",
  "D404",
  "D405",
  "D406",
  "D407",
  "D408",
  "D409",
  "D410",
  "D411",
  "D412",
  "D413",
  "D414",
  "D416",
  "D417",
  "RET504",
  "S101",
]
